refactor: organize widget feature files
