+++ presentation/widget/src/main/java/com/hilingual/presentation/widget/streak/StreakUiState.kt
@@ -1,0 +1,12 @@
+package com.hilingual.presentation.widget.streak
+
+internal data class StreakUiState(
+    val isLoggedIn: Boolean,
+    val streakDays: Int,
+    val recentDays: List<StreakDay>,
+)
+
+internal data class StreakDay(
+    val label: String,
+    val isWritten: Boolean,
+)
