+++ presentation/widget/src/main/java/com/hilingual/presentation/widget/streak/StreakWidgetReceiver.kt
@@ -1,0 +1,8 @@
+package com.hilingual.presentation.widget.streak
+
+import androidx.glance.appwidget.GlanceAppWidget
+import androidx.glance.appwidget.GlanceAppWidgetReceiver
+
+class StreakWidgetReceiver : GlanceAppWidgetReceiver() {
+    override val glanceAppWidget: GlanceAppWidget = StreakWidget()
+}
