+++ presentation/widget/src/main/java/com/hilingual/presentation/widget/topic/RecommendedTopicUiState.kt
@@ -1,0 +1,13 @@
+package com.hilingual.presentation.widget.topic
+
+internal data class RecommendedTopicUiState(
+    val dateLabel: String,
+    val topic: String?,
+    val writingStatus: WritingStatus,
+    val remainingHours: Int?,
+)
+
+internal enum class WritingStatus {
+    UNWRITTEN,
+    WRITTEN,
+}
