+++ presentation/widget/src/main/java/com/hilingual/presentation/widget/topic/RecommendedTopicWidgetReceiver.kt
@@ -1,0 +1,8 @@
+package com.hilingual.presentation.widget.topic
+
+import androidx.glance.appwidget.GlanceAppWidget
+import androidx.glance.appwidget.GlanceAppWidgetReceiver
+
+class RecommendedTopicWidgetReceiver : GlanceAppWidgetReceiver() {
+    override val glanceAppWidget: GlanceAppWidget = RecommendedTopicWidget()
+}
